Apply changes to renamed file after rebase
