+++ frontend/src/modules/SimulationTimeSeries/settings/atoms/queryAtoms.ts
@@ -21,7 +21,7 @@ export const vectorListQueriesAtom = atomWithQueries((get) => {
                         query: {
                             case_uuid: ensembleIdent.getCaseUuid(),
                             ensemble_name: ensembleIdent.getEnsembleName(),
-                            include_cumulative_vectors: true,
+                            include_derived_vectors: true,
                         },
                         throwOnError: true,
                     });
